+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+* Return correct platform event type when buffering
+
 ## 1.0.0
 
 * Initial release.